+++ ft_printf/nbr_management/ft_uitoa_ext_buffer_up_10.c
@@ -6,18 +6,18 @@
 /*   By: scarboni <scarboni@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/26 08:00:02 by scarboni          #+#    #+#             */
-/*   Updated: 2020/08/19 16:44:18 by scarboni         ###   ########.fr       */
+/*   Updated: 2020/08/24 16:33:24 by scarboni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../ft_printf.h"
 
-int	ft_uitoa_ext_buffer_up_10(unsigned int nbr, char *buffer, int base,
+int	ft_uitoa_ext_buffer_up_10(unsigned long nbr, char *buffer, int base,
 	int faux_chiffre)
 {
-	unsigned int	i;
-	unsigned int	tmp;
-	unsigned int	len;
+	unsigned long	i;
+	unsigned long	tmp;
+	unsigned long	len;
 
 	if (!buffer)
 		return (-EXIT_FAILURE);
