+++ ft_printf/nbr_management/uitoa_len.c
@@ -6,15 +6,15 @@
 /*   By: scarboni <scarboni@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/26 08:00:02 by scarboni          #+#    #+#             */
-/*   Updated: 2020/08/15 14:49:48 by scarboni         ###   ########.fr       */
+/*   Updated: 2020/08/24 16:33:24 by scarboni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../ft_printf.h"
 
-unsigned int	uitoa_len(unsigned int nb, int base)
+unsigned long	uitoa_len(unsigned long nb, int base)
 {
-	unsigned int	p;
+	unsigned long	p;
 
 	p = 1;
 	if (nb < 0)
